added helpfulquestion and questionsearch components, additional styling
