+++ client/src/components/Questions/Individualquestion.jsx
@@ -2,39 +2,17 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 
 export default function IndividualQuestion() {
-  let [helpfulQuestionYes, setHelpfulQuestionYes] = useState(0);
   let [helpfulAnswerYes, setHelpfulAnswerYes] = useState(0);
   const [report, setReport] = useState(0);
 
-  function handleQuestionOnClick() {
-    setHelpfulQuestionYes(helpfulQuestionYes += 1);
-  }
   function handleAnswerOnClick() {
     setHelpfulAnswerYes(helpfulAnswerYes += 1);
-  }
-  function getReq() {
-    axios.get('/qa/questions', { params: { product_id: 44388 } })
-      .then((response) => {
-        console.log('response', response);
-      })
-      .catch();
   }
 
   return (
     <div>
       <div>Q:Who what which where why whether how?</div>
       <div>A:One two threee four five six seven</div>
-      <div>
-        <span>Helpful? </span>
-        <span className="inline" onClick={handleQuestionOnClick}>
-          Yes
-          (
-          {helpfulQuestionYes}
-          )
-        </span>
-        <span className="inline"> | </span>
-        <span className="inline" onClick={getReq}> Add answer </span>
-      </div>
       <div>
         <span>by User1234, Octotober 30th, 2021 |</span>
         <span> | </span>
