+++ client/src/components/Questions/QuestionSearch.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+export default function SearchQuestion() {
+  return (
+    <div>
+      <input
+        id="searchbar"
+        type="text"
+        placeholder="HAVE A QUESTION? SEARCH FOR ANSWERS..."
+      />
+
+    </div>
+  );
+}
