+++ client/src/components/Questions/Questions.jsx
@@ -1,15 +1,19 @@
 import React, { useState, useEffect, useContext } from 'react';
 import IndividualQuestion from './Individualquestion.jsx';
+import SearchQuestion from './QuestionSearch.jsx';
+import HelpfulQuestion from './HelpfulQuestion.jsx';
 
 export default function Questions() {
   // declare state variables here
 
   return (
     <div>
-      <div className="container">
+      <div className="questioncontainer">
         {' '}
-        <span>QUESTIONS & ANSWERS</span>
+        <div id="qandaheader">QUESTIONS & ANSWERS</div>
+        <div><SearchQuestion /></div>
         <div><IndividualQuestion /></div>
+        <div><HelpfulQuestion /></div>
       </div>
     </div>
   );
